refactor: adjust imports and filenames
