+++ AngularStandalone/src/app/markdown-document/search/view-type-options.model.ts
