+++ AngularStandalone/src/app/shared/lists/expansion-document-list/categorized-search-result.model.ts
