+++ AngularStandalone/src/app/shared/loading-bar/loading-bar.component.ts
@@ -1,4 +1,5 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
+
 import { LoadingBarService } from './loading-bar.service';
 
 @Component({
