+++ AngularStandalone/src/app/shared/services/markdown-document-title-strategy.service.ts
@@ -2,6 +2,7 @@ import { Injectable, OnDestroy } from '@angular/core';
 import { RouterStateSnapshot, TitleStrategy } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { Subject, takeUntil } from 'rxjs';
+
 import { selectDocumentTitle } from 'src/app/markdown-document/store/markdown-document.selectors';
 import { environment } from 'src/environments/environment';
